Current Jacket Man Textures
This new and improved Jacket Man has shinier clothing and textures that stand out more and are reflective. Making it more suited for our game which revolves around lighting. The other upside to this character compared to old is that this character takes up 95% less storage space in the machine compared to before. Dark and far images are used to get a sense of what character looks like in game.